pre resolve input components during field enrichment to simplify field component
