+++ lib/interviewer/containers/TanStackForm/Field.tsx
@@ -1,43 +1,16 @@
-import {
-  type ComponentType,
-  ComponentTypes,
-} from '@codaco/protocol-validation';
 import { type VariableValue } from '@codaco/shared-consts';
-import { get } from 'es-toolkit/compat';
-import { useMemo } from 'react';
-import * as Fields from '~/lib/ui/components/Fields';
 import { useFieldContext } from '../../utils/formContexts';
 import type { FieldProps, InputComponentProps } from './types';
 
-const ComponentTypeNotFound = (componentType: string) => {
-  const NotFoundComponent = () => (
-    <div>Input component &quot;{componentType}&quot; not found.</div>
-  );
-  NotFoundComponent.displayName = `ComponentTypeNotFound(${componentType})`;
-  return NotFoundComponent;
-};
-
-/*
-* Returns the named field component, if no matching one is found
-or else it just returns a text input
-* @param {object} field The properties handed down from the protocol form
-*/
-
-const getInputComponent = (componentType: ComponentType = 'Text') => {
-  const def = get(ComponentTypes, componentType);
-  return get(Fields, def, ComponentTypeNotFound(componentType));
-};
-
 const Field = ({ field, autoFocus }: FieldProps) => {
   const fieldContext = useFieldContext();
-  const { fieldLabel, name, component, options, type, parameters, label } =
+  const { fieldLabel, name, component, options, type, parameters, label, Component } =
     field;
 
-  const InputComponent = useMemo<React.ComponentType<InputComponentProps>>(
-    () =>
-      getInputComponent(component) as React.ComponentType<InputComponentProps>,
-    [component],
-  );
+  // Use pre-resolved component if available, otherwise fallback to runtime resolution
+  if (!Component) {
+    throw new Error(`Component not resolved for field: ${name}`);
+  }
 
   const inputProps: InputComponentProps = {
     input: {
@@ -63,7 +36,7 @@ const Field = ({ field, autoFocus }: FieldProps) => {
     type,
   };
 
-  return <InputComponent {...inputProps} />;
+  return <Component {...inputProps} />;
 };
 
 export default Field;
