+++ lib/interviewer/hooks/TanstackForm/useFormData.ts
@@ -1,7 +1,13 @@
+import {
+  type ComponentType,
+  ComponentTypes,
+} from '@codaco/protocol-validation';
 import { type VariableValue } from '@codaco/shared-consts';
-import { useMemo } from 'react';
+import { get } from 'es-toolkit/compat';
+import React, { useMemo } from 'react';
 import { useStore as useReduxStore, useSelector } from 'react-redux';
-import type { FieldType, FormField } from '../../containers/TanStackForm/types';
+import * as Fields from '~/lib/ui/components/Fields';
+import type { FieldType, FormField, InputComponentProps } from '../../containers/TanStackForm/types';
 import { makeEnrichFieldsWithCodebookMetadata } from '../../selectors/forms';
 import { getCodebookVariablesForSubjectType } from '../../selectors/protocol';
 import {
@@ -11,6 +17,19 @@ import {
 import { type AppStore } from '../../store';
 import { getTanStackNativeValidators } from '../../utils/field-validation';
 import { type ValidationContext } from '../../utils/formContexts';
+
+const ComponentTypeNotFound = (componentType: string) => {
+  const NotFoundComponent = () => {
+    return React.createElement('div', {}, `Input component "${componentType}" not found.`);
+  };
+  NotFoundComponent.displayName = `ComponentTypeNotFound(${componentType})`;
+  return NotFoundComponent;
+};
+
+const getInputComponent = (componentType: ComponentType = 'Text') => {
+  const def = get(ComponentTypes, componentType);
+  return get(Fields, def, ComponentTypeNotFound(componentType));
+};
 
 export type UseFormDataReturn = {
   enrichedFields: FieldType[];
@@ -69,9 +88,13 @@ export const useFormData = ({
       // Build default values
       defaults[field.name] = initialValues?.[field.name] ?? '';
 
+      // Pre-resolve component to avoid runtime lookups
+      const Component = getInputComponent(field.component) as React.ComponentType<InputComponentProps>;
+
       // Return field with additional props
       return {
         ...field,
+        Component,
         isFirst: autoFocus && index === 0,
         validators: getTanStackNativeValidators(
           field.validation ?? {},
